Adding support for filtering on sprint
This also necessitated a move towards more of the JIRA Software API
(i.e. the `/rest/agile` endpoint, rather than the `/rest/api` endpoint).
This means that it's harder to filter on Projects, so I've switched to
Boards.

For the most part, the UI is the same, but the guts are changed.

Signed-off-by: Dave Henderson <dhenderson@gmail.com>